+++ Quan_li_ban_hoa/nbproject/project.properties
@@ -33,10 +33,12 @@ dist.javadoc.dir=${dist.dir}/javadoc
 dist.jlink.dir=${dist.dir}/jlink
 dist.jlink.output=${dist.jlink.dir}/Quan_li_ban_hoa
 excludes=
+file.reference.mssql-jdbc-12.10.0.jre11.jar=C:\\Users\\TOSHIBA\\Downloads\\sqljdbc_12.10.0.0_enu\\sqljdbc_12.10\\enu\\jars\\mssql-jdbc-12.10.0.jre11.jar
 includes=**
 jar.compress=false
 javac.classpath=\
-    ${libs.absolutelayout.classpath}
+    ${libs.absolutelayout.classpath}:\
+    ${file.reference.mssql-jdbc-12.10.0.jre11.jar}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
