+++ Quan_li_ban_hoa/src/QLBH_MVC/Khach_Hang_Chi_Tiet.java
@@ -8,12 +8,12 @@ package QLBH_MVC;
  *
  * @author TOSHIBA
  */
-public class Customer_Details extends javax.swing.JFrame {
+public class Khach_Hang_Chi_Tiet extends javax.swing.JFrame {
 
     /**
      * Creates new form Customer_Details
      */
-    public Customer_Details() {
+    public Khach_Hang_Chi_Tiet() {
         initComponents();
     }
 
@@ -219,20 +219,21 @@ public class Customer_Details extends javax.swing.JFrame {
                 }
             }
         } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(Customer_Details.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(Khach_Hang_Chi_Tiet.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(Customer_Details.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(Khach_Hang_Chi_Tiet.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(Customer_Details.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(Khach_Hang_Chi_Tiet.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(Customer_Details.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(Khach_Hang_Chi_Tiet.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
+        //</editor-fold>
         //</editor-fold>
 
         /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
             public void run() {
-                new Customer_Details().setVisible(true);
+                new Khach_Hang_Chi_Tiet().setVisible(true);
             }
         });
     }
